# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Analysis mode
 */
export type AnalyzeChartBodyMode =
  (typeof AnalyzeChartBodyMode)[keyof typeof AnalyzeChartBodyMode];

export const AnalyzeChartBodyMode = {
  SOLO: "SOLO",
  OCTOGONE: "OCTOGONE",
} as const;

export interface AnalyzeChartBody {
  /** Base64-encoded image of the TradingView chart */
  imageBase64: string;
  /** Asset name (e.g. GC1!, NQ1!, BTC/USD) */
  asset?: string;
  /** Chart timeframe (e.g. 15m, 1h) */
  timeframe?: string;
  /** Analysis mode */
  mode?: AnalyzeChartBodyMode;
}

/**
 * Trading signal
 */
export type AnalysisResultSignal =
  (typeof AnalysisResultSignal)[keyof typeof AnalysisResultSignal];

export const AnalysisResultSignal = {
  BUY: "BUY",
  SELL: "SELL",
  WAIT: "WAIT",
  NO_SIGNAL: "NO_SIGNAL",
} as const;

/**
 * Analysis mode
 */
export type AnalysisResultMode =
  (typeof AnalysisResultMode)[keyof typeof AnalysisResultMode];

export const AnalysisResultMode = {
  SOLO: "SOLO",
  OCTOGONE: "OCTOGONE",
} as const;

export interface EnergyAnalysis {
  /** Barres bleues / Barres grises / Absent / Accélération */
  state: string;
  /** Faible / Modérée / Forte */
  intensity: string;
  /** Haussière / Baissière / Aucune */
  divergence: string;
  /** Description of oscillator */
  detail: string;
}

export interface DirectionAnalysis {
  /** HAUSSIÈRE / BAISSIÈRE / NEUTRE */
  trend: string;
  /** Verts / Rouges / En retournement */
  points: string;
  /** En expansion / En contraction / Stables */
  bands: string;
  /** OUI / NON */
  springEffect: string;
}

export interface PivotsAnalysis {
  /**
   * Current price level
   * @nullable
   */
  currentPrice?: string | null;
  /** DO / RÉ / MI / FA / SOL / LA / SI */
  nearestPivot: string;
  /** Support / Résistance */
  pivotType: string;
  /** Next pivot level */
  nextPivot: string;
}

/**
 * @nullable
 */
export type TradeDetails = {
  /** Entry level or pivot */
  entry?: string;
  /** Stop loss level */
  stopLoss?: string;
  /** Take profit 1 level */
  tp1?: string;
  /** Take profit 2 level */
  tp2?: string;
} | null;

export interface AnalysisResult {
  /** Trading signal */
  signal: AnalysisResultSignal;
  /**
   * Asset name
   * @nullable
   */
  asset?: string | null;
  /**
   * Chart timeframe
   * @nullable
   */
  timeframe?: string | null;
  energy: EnergyAnalysis;
  direction: DirectionAnalysis;
  pivots: PivotsAnalysis;
  trade?: TradeDetails | null;
  /** Missing conditions for signal confirmation */
  missingConditions?: string;
  /** Brief synthesis of the analysis */
  synthesis: string;
  /** Full raw analysis text from AI */
  rawAnalysis: string;
  analyzedAt: string;
  /** Analysis mode */
  mode: AnalysisResultMode;
  /** Whether the AI-provided TP violated Belkhayate pivot rules */
  tpMismatch: boolean;
  /**
   * Detail of the TP rule violation when tpMismatch is true
   * @nullable
   */
  tpMismatchDetail?: string | null;
  /** Whether the TP was automatically corrected by the system */
  tpCorrected: boolean;
}

/**
 * Analysis mode
 * @nullable
 */
export type AnalysisRecordMode =
  | (typeof AnalysisRecordMode)[keyof typeof AnalysisRecordMode]
  | null;

export const AnalysisRecordMode = {
  SOLO: "SOLO",
  OCTOGONE: "OCTOGONE",
} as const;

export interface AnalysisRecord {
  id: number;
  signal: string;
  /** @nullable */
  asset?: string | null;
  /** @nullable */
  timeframe?: string | null;
  synthesis: string;
  rawAnalysis: string;
  /** @nullable */
  imageBase64?: string | null;
  energyState: string;
  directionTrend: string;
  nearestPivot: string;
  /** @nullable */
  entry?: string | null;
  /** @nullable */
  stopLoss?: string | null;
  /** @nullable */
  tp1?: string | null;
  /** @nullable */
  tp2?: string | null;
  /**
   * Analysis mode
   * @nullable
   */
  mode?: AnalysisRecordMode;
  /**
   * Octogone session identifier — shared by all markets in one analysis run
   * @nullable
   */
  sessionId?: string | null;
  createdAt: string;
}

export interface OctogoneMacroRegime {
  regime: string;
  count: number;
}

export interface AnalysisStats {
  total: number;
  buySignals: number;
  sellSignals: number;
  waitSignals: number;
  noSignals: number;
  todayCount: number;
  /** Total analyses in SOLO mode */
  soloTotal: number;
  /** Total analyses in OCTOGONE mode */
  octogoneTotal: number;
  /** Regime breakdown for OCTOGONE analyses */
  octogoneMacroRegimes: OctogoneMacroRegime[];
  /** Total BUY/SELL analyses where the AI TP violated pivot rules */
  tpMismatchTotal: number;
  /** Total analyses where the TP was automatically corrected */
  tpCorrectedTotal: number;
}

export interface ApiError {
  error: string;
}