# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Analyze a TradingView chart screenshot using Belkhayate method
 */
export const analyzeChartBodyTimeframeDefault = `15m`;
export const analyzeChartBodyModeDefault = `SOLO`;

export const AnalyzeChartBody = zod.object({
  imageBase64: zod
    .string()
    .describe("Base64-encoded image of the TradingView chart"),
  asset: zod
    .string()
    .optional()
    .describe("Asset name (e.g. GC1!, NQ1!, BTC\/USD)"),
  timeframe: zod
    .string()
    .default(analyzeChartBodyTimeframeDefault)
    .describe("Chart timeframe (e.g. 15m, 1h)"),
  mode: zod
    .enum(["SOLO", "OCTOGONE"])
    .default(analyzeChartBodyModeDefault)
    .describe("Analysis mode"),
});

export const AnalyzeChartResponse = zod.object({
  signal: zod
    .enum(["BUY", "SELL", "WAIT", "NO_SIGNAL"])
    .describe("Trading signal"),
  asset: zod.string().nullish().describe("Asset name"),
  timeframe: zod.string().nullish().describe("Chart timeframe"),
  energy: zod.object({
    state: zod
      .string()
      .describe("Barres bleues \/ Barres grises \/ Absent \/ Accélération"),
    intensity: zod.string().describe("Faible \/ Modérée \/ Forte"),
    divergence: zod.string().describe("Haussière \/ Baissière \/ Aucune"),
    detail: zod.string().describe("Description of oscillator"),
  }),
  direction: zod.object({
    trend: zod.string().describe("HAUSSIÈRE \/ BAISSIÈRE \/ NEUTRE"),
    points: zod.string().describe("Verts \/ Rouges \/ En retournement"),
    bands: zod.string().describe("En expansion \/ En contraction \/ Stables"),
    springEffect: zod.string().describe("OUI \/ NON"),
  }),
  pivots: zod.object({
    currentPrice: zod.string().nullish().describe("Current price level"),
    nearestPivot: zod
      .string()
      .describe("DO \/ RÉ \/ MI \/ FA \/ SOL \/ LA \/ SI"),
    pivotType: zod.string().describe("Support \/ Résistance"),
    nextPivot: zod.string().describe("Next pivot level"),
  }),
  trade: zod
    .object({
      entry: zod.string().optional().describe("Entry level or pivot"),
      stopLoss: zod.string().optional().describe("Stop loss level"),
      tp1: zod.string().optional().describe("Take profit 1 level"),
      tp2: zod.string().optional().describe("Take profit 2 level"),
    })
    .nullish(),
  missingConditions: zod
    .string()
    .optional()
    .describe("Missing conditions for signal confirmation"),
  synthesis: zod.string().describe("Brief synthesis of the analysis"),
  rawAnalysis: zod.string().describe("Full raw analysis text from AI"),
  analyzedAt: zod.coerce.date(),
  mode: zod.enum(["SOLO", "OCTOGONE"]).describe("Analysis mode"),
  tpMismatch: zod
    .boolean()
    .describe("Whether the AI-provided TP violated Belkhayate pivot rules"),
  tpMismatchDetail: zod
    .string()
    .nullish()
    .describe("Detail of the TP rule violation when tpMismatch is true"),
  tpCorrected: zod
    .boolean()
    .describe("Whether the TP was automatically corrected by the system"),
});

/**
 * @summary List past analyses
 */
export const ListAnalysisHistoryResponseItem = zod.object({
  id: zod.number(),
  signal: zod.string(),
  asset: zod.string().nullish(),
  timeframe: zod.string().nullish(),
  synthesis: zod.string(),
  rawAnalysis: zod.string(),
  imageBase64: zod.string().nullish(),
  energyState: zod.string(),
  directionTrend: zod.string(),
  nearestPivot: zod.string(),
  entry: zod.string().nullish(),
  stopLoss: zod.string().nullish(),
  tp1: zod.string().nullish(),
  tp2: zod.string().nullish(),
  mode: zod
    .union([zod.literal("SOLO"), zod.literal("OCTOGONE"), zod.literal(null)])
    .nullish()
    .describe("Analysis mode"),
  sessionId: zod
    .string()
    .nullish()
    .describe(
      "Octogone session identifier — shared by all markets in one analysis run",
    ),
  createdAt: zod.coerce.date(),
});
export const ListAnalysisHistoryResponse = zod.array(
  ListAnalysisHistoryResponseItem,
);

/**
 * @summary Get a specific analysis record
 */
export const GetAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnalysisResponse = zod.object({
  id: zod.number(),
  signal: zod.string(),
  asset: zod.string().nullish(),
  timeframe: zod.string().nullish(),
  synthesis: zod.string(),
  rawAnalysis: zod.string(),
  imageBase64: zod.string().nullish(),
  energyState: zod.string(),
  directionTrend: zod.string(),
  nearestPivot: zod.string(),
  entry: zod.string().nullish(),
  stopLoss: zod.string().nullish(),
  tp1: zod.string().nullish(),
  tp2: zod.string().nullish(),
  mode: zod
    .union([zod.literal("SOLO"), zod.literal("OCTOGONE"), zod.literal(null)])
    .nullish()
    .describe("Analysis mode"),
  sessionId: zod
    .string()
    .nullish()
    .describe(
      "Octogone session identifier — shared by all markets in one analysis run",
    ),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete an analysis record
 */
export const DeleteAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get statistics about past analyses
 */
export const GetAnalysisStatsResponse = zod.object({
  total: zod.number(),
  buySignals: zod.number(),
  sellSignals: zod.number(),
  waitSignals: zod.number(),
  noSignals: zod.number(),
  todayCount: zod.number(),
  soloTotal: zod.number().describe("Total analyses in SOLO mode"),
  octogoneTotal: zod.number().describe("Total analyses in OCTOGONE mode"),
  octogoneMacroRegimes: zod
    .array(
      zod.object({
        regime: zod.string(),
        count: zod.number(),
      }),
    )
    .describe("Regime breakdown for OCTOGONE analyses"),
  tpMismatchTotal: zod
    .number()
    .describe("Total BUY/SELL analyses where the AI TP violated pivot rules"),
  tpCorrectedTotal: zod
    .number()
    .describe("Total analyses where the TP was automatically corrected"),
});