# lib/api-zod/src/generated/types/analysisRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRecordMode } from "./analysisRecordMode";

export interface AnalysisRecord {
  id: number;
  signal: string;
  /** @nullable */
  asset?: string | null;
  /** @nullable */
  timeframe?: string | null;
  synthesis: string;
  rawAnalysis: string;
  /** @nullable */
  imageBase64?: string | null;
  energyState: string;
  directionTrend: string;
  nearestPivot: string;
  /** @nullable */
  entry?: string | null;
  /** @nullable */
  stopLoss?: string | null;
  /** @nullable */
  tp1?: string | null;
  /** @nullable */
  tp2?: string | null;
  /**
   * Analysis mode
   * @nullable
   */
  mode?: AnalysisRecordMode;
  /**
   * Octogone session identifier — shared by all markets in one analysis run
   * @nullable
   */
  sessionId?: string | null;
  createdAt: Date;
}