# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultMode } from "./analysisResultMode";
import type { AnalysisResultSignal } from "./analysisResultSignal";
import type { DirectionAnalysis } from "./directionAnalysis";
import type { EnergyAnalysis } from "./energyAnalysis";
import type { PivotsAnalysis } from "./pivotsAnalysis";
import type { TradeDetails } from "./tradeDetails";

export interface AnalysisResult {
  /** Trading signal */
  signal: AnalysisResultSignal;
  /**
   * Asset name
   * @nullable
   */
  asset?: string | null;
  /**
   * Chart timeframe
   * @nullable
   */
  timeframe?: string | null;
  energy: EnergyAnalysis;
  direction: DirectionAnalysis;
  pivots: PivotsAnalysis;
  trade?: TradeDetails | null;
  /** Missing conditions for signal confirmation */
  missingConditions?: string;
  /** Brief synthesis of the analysis */
  synthesis: string;
  /** Full raw analysis text from AI */
  rawAnalysis: string;
  analyzedAt: Date;
  /** Analysis mode */
  mode: AnalysisResultMode;
}