# lib/api-zod/src/generated/types/analysisStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OctogoneMacroRegime } from "./octogoneMacroRegime";

export interface AnalysisStats {
  total: number;
  buySignals: number;
  sellSignals: number;
  waitSignals: number;
  noSignals: number;
  todayCount: number;
  /** Total analyses in SOLO mode */
  soloTotal: number;
  /** Total analyses in OCTOGONE mode */
  octogoneTotal: number;
  /** Regime breakdown for OCTOGONE analyses */
  octogoneMacroRegimes: OctogoneMacroRegime[];
}