# lib/api-zod/src/generated/types/analyzeChartBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeChartBodyMode } from "./analyzeChartBodyMode";

export interface AnalyzeChartBody {
  /** Base64-encoded image of the TradingView chart */
  imageBase64: string;
  /** Asset name (e.g. GC1!, NQ1!, BTC/USD) */
  asset?: string;
  /** Chart timeframe (e.g. 15m, 1h) */
  timeframe?: string;
  /** Analysis mode */
  mode?: AnalyzeChartBodyMode;
}