# lib/api-zod/src/generated/types/directionAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DirectionAnalysis {
  /** HAUSSIÈRE / BAISSIÈRE / NEUTRE */
  trend: string;
  /** Verts / Rouges / En retournement */
  points: string;
  /** En expansion / En contraction / Stables */
  bands: string;
  /** OUI / NON */
  springEffect: string;
}