# lib/api-zod/src/generated/types/energyAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyAnalysis {
  /** Barres bleues / Barres grises / Absent / Accélération */
  state: string;
  /** Faible / Modérée / Forte */
  intensity: string;
  /** Haussière / Baissière / Aucune */
  divergence: string;
  /** Description of oscillator */
  detail: string;
}