# lib/api-zod/src/generated/types/tradeDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * @nullable
 */
export type TradeDetails = {
  /** Entry level or pivot */
  entry?: string;
  /** Stop loss level */
  stopLoss?: string;
  /** Take profit 1 level */
  tp1?: string;
  /** Take profit 2 level */
  tp2?: string;
} | null;